# fix module source cache invalidation + cleanup api