--- a/cmd/codegen/generator/typescript/generator.go
+++ b/cmd/codegen/generator/typescript/generator.go
@@ -63,7 +63,7 @@ func (g *TypeScriptGenerator) Generate(_ context.Context, schema *introspection.
 
 	target := ClientGenFile
 	if g.Config.ModuleName != "" {
-		target = filepath.Join(g.Config.ModuleContextPath, "sdk/src/api", ClientGenFile)
+		target = filepath.Join(g.Config.ModuleSourcePath, "sdk/src/api", ClientGenFile)
 	}
 	if err := mfs.MkdirAll(filepath.Dir(target), 0700); err != nil {
 		return nil, err