--- a/cmd/dagger/modconf.graphql
+++ b/cmd/dagger/modconf.graphql
@@ -3,9 +3,7 @@ query ModuleConfig($source: ModuleSourceID!) {
     asString
     module: asModule {
       name
-      initialize {
-        description
-      }
+      description
       dependencies {
         name
         description