--- a/cmd/dagger/modconf.graphql
+++ b/cmd/dagger/modconf.graphql
@@ -1,6 +1,7 @@
 query ModuleConfig($source: ModuleSourceID!) {
   source: loadModuleSourceFromID(id: $source) {
     asString
+    kind
     module: asModule {
       name
       description