… .changes/unreleased/Breaking-20250213-202756.yaml
… .github/workflows/daggerverse-preview.gen.yml
… .github/workflows/docs.gen.yml
… .github/workflows/engine-cli.gen.yml
… .github/workflows/github.gen.yml
… .github/workflows/helm.gen.yml
… .github/workflows/sdks.gen.yml
… cmd/dagger/module.go
… cmd/dagger/module_inspect.go
… cmd/dagger/shell_commands.go
… cmd/dagger/shell_exec.go
… core/ids.go
… core/integration/legacy_test.go
… core/integration/module_call_test.go
… core/integration/module_cli_test.go
… core/integration/module_config_test.go
… core/integration/module_go_test.go
… core/integration/module_python_test.go
… core/integration/module_test.go
… core/integration/module_typescript_test.go
… core/integration/provision_test.go
… core/integration/suite_test.go
… core/modfunc.go
… core/module.go
… core/modules/config.go
… core/modulesource.go
… core/modulesource_test.go
… core/object.go
… core/schema/cache.go
… core/schema/coremod.go