Goals here are:
* Fix various sources of excessive module building cache invalidation
   * changing completely unrelated source files in context
   * building different git commits that have no changes in the module source
   * etc.
* Use the filesync improvements + new caching features in #9204 to cleanup tons of cruft from the API (ResolveFromCaller, Module.initialize, etc.)

So far the results are looking good.

Comparing the runtime of the `load module` step for the dagger repo:
1. Completely uncached - unchanged (expected)
2. Cached and with random change to engine source code
   * Previous: ~30s
   * Now: ~3s
3. Cached and with random change to dagger module source code
   * Previous: ~10s
   * Now: ~5s

The difference in `2.` above of 10x speedup should make our day to day development quite a bit less painful; *a lot* less waiting around an extra 30s every time any change is made.

In terms of the simplifications, the net deletion of over 6k lines of code is indicative of success so far there 😄 I think that diff will actually increase with more cleanup.

---

This will have some breaking changes in the current plans:
* dagger.json's `include` field will now be co-opted to only need to be set when a module imports code from *local* non-daggerized packages. It should never need to be set for any other case.
   * I will remove `exclude` entirely since it's redundant with `!` include patterns
* `views` are currently gonna be removed since they don't serve much of a purpose anymore thanks to contextual args

--- 

Have basic tests passing, TODOs remaining:
- [ ] get all integ tests passing, anticipating a lot of corner case squashing from stuff missed during refactoring
- [ ] dedupe a bunch of copy-pasta chunks of code
- [ ] cleanup various TODO comments (fix dagger install telemetry, etc.)
- [ ] write up breaking change 
- [ ] remove more old cruft like RequiredPaths from sdks (now unused but still laying around)