@helderco 2025-02-14
Disabling find up isn't working for local sources:
```
⋈ .core | module-source sdk/cue --disable-find-up | as-string
sdk/cue
⋈ .core | module-source github.com/dagger/dagger/sdk/cue --disable-find-up | as-string
Error: input: moduleSource git module source "github.com/dagger/dagger/sdk/cue" does not contain a dagger config file
```