@helderco 2025-02-14
Oh, git sources never do findUp:
```
⋈ .core | module-source github.com/dagger/dagger/sdk/cue | as-string
Error: input: moduleSource git module source "github.com/dagger/dagger/sdk/cue" does not contain a dagger config file
```
Can we make it do it?