@sipsma 2025-02-15
> Oh, git sources never do findUp:

Yeah for sure, added that

> I expected sdk/php here:

Changed this a bit, let me know what you think. I changed it so that the return value for local sources is the full absolute path to the source root. That makes sense in that it's consistent for a given host and not tied to the caller's cwd

However, this did break something in shell help output in that what might have previously been e.g. `Usage: .` now be `Usage: /the/full/abs/path`. So I just did some special case handling in the shell client specifically to handle that and make it a relative path for nicer looking output there (back to `.` in the example before): https://github.com/dagger/dagger/pull/9505/commits/78cc5590ddcfa10e630fc4224146a6f237441dbf