@helderco 2025-02-15
> > Oh, git sources never do findUp:
> 
> Yeah for sure, added that

Nice, that makes things simpler. 👍

> > I expected sdk/php here:
> 
> Changed this a bit, let me know what you think. I changed it so that the return value for local sources is the full absolute path to the source root. That makes sense in that it's consistent for a given host and not tied to the caller's cwd

Yeah, that's what I need, both local and git return a full ref. It's easy from the CLI to make it relative to either the context or `os.Getwd()` if I need to.

> However, this did break something in shell help output in that what might have previously been e.g. `Usage: .` now be `Usage: /the/full/abs/path`. So I just did some special case handling in the shell client specifically to handle that and make it a relative path for nicer looking output there (back to `.` in the example before): [78cc559](https://github.com/dagger/dagger/commit/78cc5590ddcfa10e630fc4224146a6f237441dbf)

Yeah, that's going to change anyway because of:
- https://github.com/dagger/dagger/pull/9511

Basically that path will be handled like in other places, which will be relative to where you've navigated to inside the context directory, rather than relative to `os.Getwd()`.

But thanks for the fix so we're not blocked on that PR.