@helderco 2025-02-15
@sipsma, I know I said I didn't need `AllowNotExists` on git sources, but would make it more consistent with local and not hard to do, right? Not a blocker, just a suggestion.