@sipsma 2025-02-18
@helderco 

> I still need for a way to get the original subpath the user used. What do you suggest?

We can just add another api for that (the original user specified path before findup, whether it was local or git).

Are you okay with adding that in your PR that's based on this one instead of including here? The rebasing and churn in this one is getting a but brutal and seems like it would be easier to tackle in yours where it's actually going to be used.