@helderco 2025-02-18
Yeah, already did that yesterday. It's on the PR.